requirements.txt
flask
redis
python-dotenv
Flask-Mail
requests
pybit
Flask-SocketIO
pandas-ta
openai-whisper
pydub
httpx